Unify alice::ActionKind and bob::ActionKind
The Accept and Decline actions should actually not be part of
Alice's actions nor Bob's actions since those roles only affect
things that happen on the blockchain. Thinking ahead, it might
easily be possible for P1 to send a swap request but declare the
role `Bob`, which means whoever receives the request is `Alice`
but still has to execute the `Accept` or `Decline` action.
