+++ application/comit_node/src/http_api/mod.rs
@@ -11,6 +11,7 @@ pub mod ledger;
 pub mod asset;
 #[macro_use]
 pub mod impl_serialize_http;
+pub mod action;
 mod problem;
 mod swap_resource;
