+++ application/comit_node/src/http_api/routes/rfc003/swap_state.rs
@@ -2,7 +2,8 @@ use crate::{
     http_api::{Http, SwapStatus},
     swap_protocols::{
         asset::Asset,
-        rfc003::{self, alice, bob, Ledger, SecretHash, Timestamp},
+        rfc003::{self, alice, bob, Ledger, SecretHash},
+        Timestamp,
     },
 };
 use serde::Serialize;
