+++ application/comit_node/src/swap_protocols/mod.rs
@@ -1,14 +1,18 @@
+pub mod actions;
 pub mod asset;
 mod dependencies;
 pub mod ledger;
-pub mod rfc003;
-
 pub mod metadata_store;
+pub mod rfc003;
+mod swap_id;
+mod timestamp;
 
 pub use self::{
     dependencies::{alice, bob, LedgerEventDependencies},
     ledger::{Ledger, LedgerKind},
     metadata_store::{InMemoryMetadataStore, Metadata, MetadataStore, RoleKind},
+    swap_id::*,
+    timestamp::Timestamp,
 };
 
 #[derive(Debug)]
@@ -16,6 +20,3 @@ pub enum SwapProtocol {
     Rfc003,
     Unknown(String),
 }
-
-mod swap_id;
-pub use self::swap_id::*;
