+++ application/comit_node/src/swap_protocols/rfc003/actions/erc20.rs
@@ -1,11 +1,13 @@
 use crate::swap_protocols::{
+    actions::ethereum::{CallContract, ContractDeploy},
     ledger::Ethereum,
-    rfc003::{ethereum, state_machine::HtlcParams, Secret, Timestamp},
+    rfc003::{state_machine::HtlcParams, Secret},
+    Timestamp,
 };
 use blockchain_contracts::ethereum::rfc003::erc20_htlc::Erc20Htlc;
 use ethereum_support::{Bytes, Erc20Token, Network};
 
-pub fn deploy_action(htlc_params: HtlcParams<Ethereum, Erc20Token>) -> ethereum::ContractDeploy {
+pub fn deploy_action(htlc_params: HtlcParams<Ethereum, Erc20Token>) -> ContractDeploy {
     htlc_params.into()
 }
 
@@ -13,11 +15,11 @@ pub fn fund_action(
     htlc_params: HtlcParams<Ethereum, Erc20Token>,
     to_erc20_contract: ethereum_support::Address,
     beta_htlc_location: ethereum_support::Address,
-) -> ethereum::CallContract {
+) -> CallContract {
     let network = htlc_params.ledger.network;
     let gas_limit = Erc20Htlc::fund_tx_gas_limit();
 
-    ethereum::CallContract {
+    CallContract {
         to: to_erc20_contract,
         data: Erc20Htlc::transfer_erc20_tx_payload(
             htlc_params.asset.quantity.0,
@@ -33,11 +35,11 @@ pub fn refund_action(
     network: Network,
     expiry: Timestamp,
     beta_htlc_location: ethereum_support::Address,
-) -> ethereum::CallContract {
+) -> CallContract {
     let data = Bytes::default();
     let gas_limit = Erc20Htlc::tx_gas_limit();
 
-    ethereum::CallContract {
+    CallContract {
         to: beta_htlc_location,
         data,
         gas_limit,
@@ -50,11 +52,11 @@ pub fn redeem_action(
     alpha_htlc_location: ethereum_support::Address,
     secret: Secret,
     network: Network,
-) -> ethereum::CallContract {
+) -> CallContract {
     let data = Bytes::from(secret.raw_secret().to_vec());
     let gas_limit = Erc20Htlc::tx_gas_limit();
 
-    ethereum::CallContract {
+    CallContract {
         to: alpha_htlc_location,
         data,
         gas_limit,
