+++ application/comit_node/src/swap_protocols/rfc003/actions/ether.rs
@@ -1,8 +1,8 @@
 use crate::swap_protocols::{
+    actions::ethereum::{CallContract, ContractDeploy},
     ledger::Ethereum,
     rfc003::{
         actions::{FundAction, RedeemAction, RefundAction},
-        ethereum,
         secret_source::SecretSource,
         state_machine::HtlcParams,
         Secret,
@@ -12,14 +12,14 @@ use blockchain_contracts::ethereum::rfc003::ether_htlc::EtherHtlc;
 use ethereum_support::{Address as EthereumAddress, Bytes, EtherQuantity};
 
 impl FundAction<Ethereum, EtherQuantity> for (Ethereum, EtherQuantity) {
-    type FundActionOutput = ethereum::ContractDeploy;
+    type FundActionOutput = ContractDeploy;
 
     fn fund_action(htlc_params: HtlcParams<Ethereum, EtherQuantity>) -> Self::FundActionOutput {
         htlc_params.into()
     }
 }
 impl RefundAction<Ethereum, EtherQuantity> for (Ethereum, EtherQuantity) {
-    type RefundActionOutput = ethereum::CallContract;
+    type RefundActionOutput = CallContract;
 
     fn refund_action(
         htlc_params: HtlcParams<Ethereum, EtherQuantity>,
@@ -29,7 +29,7 @@ impl RefundAction<Ethereum, EtherQuantity> for (Ethereum, EtherQuantity) {
         let data = Bytes::default();
         let gas_limit = EtherHtlc::tx_gas_limit();
 
-        ethereum::CallContract {
+        CallContract {
             to: htlc_location,
             data,
             gas_limit,
@@ -39,7 +39,7 @@ impl RefundAction<Ethereum, EtherQuantity> for (Ethereum, EtherQuantity) {
     }
 }
 impl RedeemAction<Ethereum, EtherQuantity> for (Ethereum, EtherQuantity) {
-    type RedeemActionOutput = ethereum::CallContract;
+    type RedeemActionOutput = CallContract;
 
     fn redeem_action(
         htlc_params: HtlcParams<Ethereum, EtherQuantity>,
@@ -50,7 +50,7 @@ impl RedeemAction<Ethereum, EtherQuantity> for (Ethereum, EtherQuantity) {
         let data = Bytes::from(secret.raw_secret().to_vec());
         let gas_limit = EtherHtlc::tx_gas_limit();
 
-        ethereum::CallContract {
+        CallContract {
             to: htlc_location,
             data,
             gas_limit,
