+++ application/comit_node/src/swap_protocols/rfc003/alice/actions/erc20.rs
@@ -1,10 +1,10 @@
 use crate::swap_protocols::{
+    actions::{ethereum, Actions},
     asset::Asset,
     ledger::Ethereum,
     rfc003::{
-        actions::{erc20, Actions, FundAction, RedeemAction, RefundAction},
+        actions::{erc20, Accept, ActionKind, Decline, FundAction, RedeemAction, RefundAction},
         alice::{self, SwapCommunication},
-        ethereum,
         state_machine::HtlcParams,
         Ledger, LedgerState,
     },
@@ -19,7 +19,9 @@ where
     (BL, BA): RedeemAction<BL, BA>,
 {
     #[allow(clippy::type_complexity)]
-    type ActionKind = alice::ActionKind<
+    type ActionKind = ActionKind<
+        Accept<Ethereum, BL>,
+        Decline<Ethereum, BL>,
         ethereum::ContractDeploy,
         ethereum::CallContract,
         <(BL, BA) as RedeemAction<BL, BA>>::RedeemActionOutput,
@@ -40,15 +42,15 @@ where
         use self::LedgerState::*;
 
         let mut actions = match alpha_state {
-            NotDeployed => vec![alice::ActionKind::Deploy(erc20::deploy_action(
+            NotDeployed => vec![ActionKind::Deploy(erc20::deploy_action(
                 HtlcParams::new_alpha_params(request, response),
             ))],
-            Deployed { htlc_location, .. } => vec![alice::ActionKind::Fund(erc20::fund_action(
+            Deployed { htlc_location, .. } => vec![ActionKind::Fund(erc20::fund_action(
                 HtlcParams::new_alpha_params(request, response),
                 request.alpha_asset.token_contract,
                 *htlc_location,
             ))],
-            Funded { htlc_location, .. } => vec![alice::ActionKind::Refund(erc20::refund_action(
+            Funded { htlc_location, .. } => vec![ActionKind::Refund(erc20::refund_action(
                 request.alpha_ledger.network,
                 request.alpha_expiry,
                 *htlc_location,
@@ -57,7 +59,7 @@ where
         };
 
         if let Funded { htlc_location, .. } = beta_state {
-            actions.push(alice::ActionKind::Redeem(<(BL, BA)>::redeem_action(
+            actions.push(ActionKind::Redeem(<(BL, BA)>::redeem_action(
                 HtlcParams::new_beta_params(request, response),
                 htlc_location.clone(),
                 &*self.secret_source,
@@ -76,7 +78,9 @@ where
     (AL, AA): RefundAction<AL, AA>,
 {
     #[allow(clippy::type_complexity)]
-    type ActionKind = alice::ActionKind<
+    type ActionKind = ActionKind<
+        Accept<AL, Ethereum>,
+        Decline<AL, Ethereum>,
         Infallible,
         <(AL, AA) as FundAction<AL, AA>>::FundActionOutput,
         ethereum::CallContract,
@@ -97,21 +101,19 @@ where
         use self::LedgerState::*;
 
         let mut actions = match alpha_state {
-            NotDeployed => vec![alice::ActionKind::Fund(<(AL, AA)>::fund_action(
+            NotDeployed => vec![ActionKind::Fund(<(AL, AA)>::fund_action(
                 HtlcParams::new_alpha_params(request, response),
             ))],
-            Funded { htlc_location, .. } => {
-                vec![alice::ActionKind::Refund(<(AL, AA)>::refund_action(
-                    HtlcParams::new_alpha_params(request, response),
-                    htlc_location.clone(),
-                    &*self.secret_source,
-                ))]
-            }
+            Funded { htlc_location, .. } => vec![ActionKind::Refund(<(AL, AA)>::refund_action(
+                HtlcParams::new_alpha_params(request, response),
+                htlc_location.clone(),
+                &*self.secret_source,
+            ))],
             _ => vec![],
         };
 
         if let Funded { htlc_location, .. } = beta_state {
-            actions.push(alice::ActionKind::Redeem(erc20::redeem_action(
+            actions.push(ActionKind::Redeem(erc20::redeem_action(
                 *htlc_location,
                 self.secret_source.secret(),
                 request.beta_ledger.network,
