+++ application/comit_node/src/swap_protocols/rfc003/alice/actions/mod.rs
@@ -1,10 +1,2 @@
 mod erc20;
 mod generic_impl;
-
-#[derive(Debug, Clone, PartialOrd, Ord, PartialEq, Eq, Hash)]
-pub enum ActionKind<Deploy, Fund, Redeem, Refund> {
-    Deploy(Deploy),
-    Fund(Fund),
-    Redeem(Redeem),
-    Refund(Refund),
-}
