--- application/comit_node/src/swap_protocols/rfc003/bitcoin/actions.rs
@@ -1,26 +1,0 @@
-use bitcoin_support::{Address, BitcoinQuantity, Network};
-use bitcoin_witness::{PrimedInput, PrimedTransaction};
-use serde::Serialize;
-
-#[derive(Debug, Clone, PartialEq, Serialize)]
-pub struct SendToAddress {
-    pub to: Address,
-    pub amount: BitcoinQuantity,
-    pub network: Network,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct SpendOutput {
-    // Remember: One man's input is another man's output!
-    pub output: PrimedInput,
-    pub network: Network,
-}
-
-impl SpendOutput {
-    pub fn spend_to(self, to_address: Address) -> PrimedTransaction {
-        PrimedTransaction {
-            inputs: vec![self.output],
-            output_address: to_address,
-        }
-    }
-}
