+++ application/comit_node/src/swap_protocols/rfc003/ethereum/mod.rs
@@ -1,9 +1,7 @@
-mod actions;
 pub mod htlc_events;
 
-pub use self::actions::*;
-
 use crate::swap_protocols::{
+    actions::ethereum::ContractDeploy,
     ledger::Ethereum,
     rfc003::{state_machine::HtlcParams, Ledger},
 };
@@ -77,5 +75,33 @@ impl From<HtlcParams<Ethereum, Erc20Token>> for Erc20Htlc {
 impl HtlcParams<Ethereum, Erc20Token> {
     pub fn bytecode(&self) -> Bytes {
         Erc20Htlc::from(self.clone()).into()
+    }
+}
+
+impl From<HtlcParams<Ethereum, EtherQuantity>> for ContractDeploy {
+    fn from(htlc_params: HtlcParams<Ethereum, EtherQuantity>) -> Self {
+        let htlc = EtherHtlc::from(htlc_params.clone());
+        let gas_limit = htlc.deployment_gas_limit();
+
+        ContractDeploy {
+            data: htlc.into(),
+            amount: htlc_params.asset,
+            gas_limit,
+            network: htlc_params.ledger.network,
+        }
+    }
+}
+
+impl From<HtlcParams<Ethereum, Erc20Token>> for ContractDeploy {
+    fn from(htlc_params: HtlcParams<Ethereum, Erc20Token>) -> Self {
+        let htlc = Erc20Htlc::from(htlc_params.clone());
+        let gas_limit = htlc.deployment_gas_limit();
+
+        ContractDeploy {
+            data: htlc.into(),
+            amount: EtherQuantity::zero(),
+            gas_limit,
+            network: htlc_params.ledger.network,
+        }
     }
 }
