+++ application/comit_node/tests/rfc003_bitcoin_htlc.rs
@@ -12,7 +12,10 @@ use bitcoin_support::{
 use bitcoin_witness::{
     PrimedInput, PrimedTransaction, UnlockParameters, Witness, SEQUENCE_ALLOW_NTIMELOCK_NO_RBF,
 };
-use comit_node::swap_protocols::rfc003::{bitcoin::Htlc, Secret, SecretHash, Timestamp};
+use comit_node::swap_protocols::{
+    rfc003::{bitcoin::Htlc, Secret, SecretHash},
+    Timestamp,
+};
 use secp256k1_support::KeyPair;
 use spectral::prelude::*;
 use std::str::FromStr;
